+++ kubernetes/config/__init__.py
@@ -14,5 +14,5 @@
 
 from .config_exception import ConfigException
 from .incluster_config import load_incluster_config
-from .kube_config import list_kube_config_contexts
-from .kube_config import load_kube_config
+from .kube_config import (list_kube_config_contexts, load_kube_config,
+                          new_client_from_config)